Decouple some definitions
We were defining one huge set of values, so there was potential for
everything to affect everything else.

This meant that small changes to one component would cause everything
to rebuild, including all of the Haskell libraries, etc.

This decouples some of that, so we build the Haskell packages on their
own, then build everything else inheriting those Haskell packages.