Import result of one derivation; not lots
We were generating a Nix file in a derivation for each sampled size, then
passing this to `import`, which caused *evaluation* to depend on a bunch of
builds. This made time and memory usage very high.

This change makes one single Nix file, imports it once, and re-uses it over and
over. This should bring down the resource requirements.