Use '<support>' in quickspecBench internals, rather than <nixpkgs>
It looks like overriding <nixpkgs> causes tipBenchmarks.tools to break: when it
tries to make a GHC environment with 'tip-lib', its Nix expression ends up
attempting to import a .tar.gz rather than a source directory.

Not sure what change sparked this off, but in any case there's no real reason to
co-opt the name <nixpkgs> when we're only referencing it from within the same
file. Switched to <support> instead, and it seems to fix that problem.