Use fixed version of asv-nix
We were taking asv-nix from nix-config, but that uses latestGit to check for
updates, which we don't really want to do in a stable, reproducible environment
like haskell-te (as much as possible, anyway!)

Now we take asv-nix from a fixed git commit, and pass in its asv dependency from
nix-config. That's OK, since the (fixed version of) nix-config we're using uses
a fixed version of asv.