"Runner" given to quickspec is a commandline, not a command
Don't stick it in quotes, since e.g. "runhaskell -XTemplateHaskell" is not the
name of a command; yet it is a valid line to give to the shell.

This is a little horrible, since e.g. spaces in arguments won't be
honoured. That's not currently a problem, so we'll cross that bridge if we come
to it.