Massively simplify benchmarking
Now that we have simple commands like 'quickspecAsts', etc., we can benchmark
these rather than having to faff around in Python, generating environments, etc.

In particular, we now provide the Python benchmarks with 2 commands per
benchmark: a "runner" which performs the exploration and spits out some results,
and an "analyser" which is given those results and spits out stats and things.

The Python just has to run these, parse the resulting JSON and pluck out the
relevant numbers.