runCmd can't write to $out
runCmd used to make a derivation, and we would readFile/import the result. This
turned out to be a bad idea, since it forced a load of stuff to build at eval
time.

We changed runCmd to instead produce a script, with all of the necessary
environment, etc. baked in, such that it could be run from within another
derivation to produce the required output. This way, we can have derivations
depending on derivations, as intended, rather than derivations depending on
strings, and those strings depending on derivations.

We also ripped a bunch of stuff out of runCmd at the same time, since it was
rather redundant (in fact, runCmd could probably be refactored away completely,
but it's doing the job for now...). It turns out that we didn't rip out enough,
since it was still trying to write to '$out' (i.e. the result of a derivation),
rather than just leaving things to stdout