Collapse all of defs.nix into one (recursive) package set
We did have a bunch of separate 'stages': getting a version of nixpkgs, using
that to get nix-config, using that to get haskellPackages, etc.

Now all of that control/data flow is worked out automatically by using
'callPackage' and 'self'. This probably makes it easier to introduce an infinite
recursion bug, but (a) that should be obvious to a test suite and (b) we no
longer have to structure the everything we do around avoiding such bugs (which
is good, since most of the time that's just irrelevant congnitive overhead)