Parameterise testData
This lets us pass in the script we want to use, e.g. to extract ASTs or convert
TIP to Haskell or whatever.

The idea is that we can define these scripts once, but use them (a) to make
subsequent data (e.g. extracting ASTs from the Haskell packages we've converted
to TIP) but also (b) in the tests of these scripts.

It is (b) which requires us to override the script, since we don't want to get
infinite loops (i.e. using a script, whose tests depend on the script's tests
passing...)