Test dlists aren't marked as hashable
They seem to be getting marked as hashable, despite GHCi failing to find a
Serialize instance for DList, and hence causing a failure during nix-eval.

Of course, DLists (difference lists) are functions, so there isn't a Serialize
instance, so they're not hashable, so GHCi and the GHC error in nix-eval are
both correct.

Somehow, they're getting marked as hashable when they're not...