Take 'stable' from allArgs
We take in 'allArgs', which may or may not contain 'stable'. We define 'args' to
be 'allArgs' with 'stable' removed (if present). 'args' is then suitable to send
in to e.g. 'import <nixpkgs>'.

Thing is, we were propagating our value of 'stable' by taking it from 'args',
falling back to 'true' if not found. But 'args' will *never* contain 'stable',
since the whole point is that it's been removed!

Take it straight from 'allArgs' now. This may expose some breakages!