Fix exception handling syntax in benchmarks
We want to catch *anything* and log it, but this is a little tricky since we can
only catch anything by doing:

    try:
      ...do stuff
    except:
      ...log here

Yet this doesn't give us a value to put in our logs. To get a value, we could do
something like:

    try:
      ...do stuff
    except Exception as e:
      ...log here

Yet this will only catch `Exception` (and any subclasses). There are certain
things in Python which *aren't* `Exception`, and we want to know about anything
which caused our benchmark/analysis stop; especially since there's a bunch of
indirection involved (asv runs benchmarks in a subprocess).

The solution is to use sys.exc_info(), which will (impurely) get info about the
last exception, which is whatever we're handling/logging.