Allow sampler to be generated from a (Nix) list of sampled names
We were giving the sampler-making-function a file path, which samples can be
read from (newline-delimited). This lets us pass in a derivation for the file
path, and hence give the sampler derivation a dependency on the samples
derivation, which Nix will sort out at build time.

Our IsaCoSy benchmark is re-using this analysis framework (because it might as
well), but in that case we already have a specific list of sampled names (a Nix
list of Nix strings). We would like to pass these in directly, rather than write
them to a file only to be read straight back.

This change allows both options, and includes an assertion to check (at eval
time) that only one of these options is being used. This way, the haskell-te
benchmarks can maintain their eval-time/build-time separation, whilst the
IsaCoSy benchmarks can avoid writing unnecessary files.