Rename variable to avoid interference
For some reason a test derivation was failing due to a 'pkg' variable getting a
value of '/nix/store/...-set-source-epoch-to-latest.sh' instead of the Haskell
package directory it should have been.

I don't know where this conflict came from (using 'nix-shell' on the test
derivation showed that '$pkg' was already '...-set-source-epoch-to-latest.sh',
before we actually try to run anything; yet the .drv file specifies the correct
value for 'pkg'. Oh well, maybe it's a setup hook or something.

In any case, this changes the variable to 'p' instead, and this seems to work.