Change the way we import nix-helpers
It looks like the contents of 'nix-helpers' gets a bit confused when we try
overriding its overrides.

This commit alters the way we call it. Now we're applying it as an overlay on to
a fresh instantiation of nixpkgs, alongside our own overlay which overrides some
things like withNix.

We store the result in 'helpers' (*not* nix-helpers, since that would be an
infinite loop), and selectively include things from it.